requirements.txt
asgiref==3.8.1
chardet==5.2.0
Django==5.1.3
django-ckeditor==6.7.2
django-cors-headers==4.6.0
django-js-asset==3.0.1
djangorestframework==3.15.2
Faker==33.3.1
mysqlclient==2.2.6
pillow==11.1.0
psycopg2-binary==2.9.10
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
reportlab==4.3.1
six==1.17.0
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.2
